hoka/src/plugins/nonebot_plugin_apscheduler/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-apscheduler"
version = "0.1.2"
description = "APScheduler Support for NoneBot2"
authors = ["yanyongyu <yanyongyu_1@126.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
apscheduler = "^3.7.0"
nonebot2 = "^2.0.0-alpha.8"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"